// UI/doctor_appointment_rms/src/features/auth/schemas/register.schema.ts
import { z } from "zod";

export const registerSchema = z.object({
  firstName: z.string().min(2, "First name is required"),
  lastName: z.string().min(2, "Last name is required"),
  email: z.string().email("Invalid email address"),
  password: z.string().min(6, "Password must be at least 6 characters"),
  role: z.string().min(1, "Role is required"),
});

export type RegisterDto = z.infer<typeof registerSchema>;